online_converter/requirements.txt
Django==2.2.2
ffmpeg==1.4
ffprobe==0.5
python-decouple==3.1
youtube-dl==2019.7.30
django-environ==0.4.5
amqp==2.5.1
mysql==0.0.2
mysqlclient==1.4.2.post1